+++ system-tests/project-fixtures/svelte/cypress/support/component.ts
@@ -1,0 +1,16 @@
+import { mount } from 'cypress/svelte'
+import '../../src/app.css'
+
+// Augment the Cypress namespace to include type definitions for
+// your custom command.
+// Alternatively, can be defined in cypress/support/component.d.ts
+// with a <reference path="./component" /> at the top of your spec.
+declare global {
+  namespace Cypress {
+    interface Chainable {
+      mount: typeof mount
+    }
+  }
+}
+
+Cypress.Commands.add('mount', mount)