+++ system-tests/projects/svelte-webpack-configured/cypress.config.ts
@@ -1,6 +1,6 @@
-const { defineConfig } = require('cypress')
+import { defineConfig } from 'cypress'
 
-module.exports = defineConfig({
+export default defineConfig({
   component: {
     experimentalSingleTabRunMode: true,
     devServer: {