+++ system-tests/projects/svelte-webpack-configured/src/mount.cy.ts
@@ -1,24 +1,25 @@
-import Counter from './Counter.svelte'
-import Context from './Context.svelte'
-import Store from './Store.svelte'
-import { messageStore } from './store'
+import Counter from './lib/Counter.svelte'
+import CounterProp from './lib/CounterProp.svelte'
+import Context from './lib/Context.svelte'
+import Store from './lib/Store.svelte'
+import { messageStore } from './lib/store'
 
 describe('Svelte mount', () => {
   it('mounts', () => {
     cy.mount(Counter)
-    cy.contains('h1', 'Count is 0')
+    cy.contains('button', 'count is 0')
   })
 
   it('reacts to state changes', () => {
     cy.mount(Counter)
-    cy.contains('h1', 'Count is 0')
+    cy.contains('button', 'count is 0')
     cy.get('button').click()
-    cy.contains('h1', 'Count is 1')
+    cy.contains('button', 'count is 1')
   })
 
   it('accepts props', () => {
-    cy.mount(Counter, { props: { count: 42 } })
-    cy.contains('h1', 'Count is 42')
+    cy.mount(CounterProp, { props: { count: 42 } })
+    cy.contains('button', 'count is 42')
   })
 
   it('accepts context', () => {
@@ -29,14 +30,6 @@ describe('Svelte mount', () => {
 
     cy.mount(Context, { context })
     cy.contains('h1', payload.msg)
-  })
-
-  it('spies on outputs', () => {
-    cy.mount(Counter).then(({ component }) => {
-      component.$on('change', cy.spy().as('changeSpy'))
-      cy.get('button').click()
-      cy.get('@changeSpy').should('have.been.called')
-    })
   })
 
   it('anchors mounted component', () => {
@@ -57,7 +50,11 @@ describe('Svelte mount', () => {
 
   context('log', () => {
     it('displays component name in mount log', () => {
-      cy.mount(Counter)
+      // svelte 5 in development mode is a function with the name data living on an internal symbol,
+      // so we are unable to calculate the name. But production builds can correctly calculate the name on the object
+      cy.mount(Counter, {
+        log: true,
+      })
 
       cy.wrap(Cypress.$(window.top.document.body)).within(() => {
         return cy
@@ -73,7 +70,7 @@ describe('Svelte mount', () => {
     })
 
     it('does not display mount log', () => {
-      cy.mount(Counter, { log: false })
+      cy.mount(Counter)
 
       cy.wrap(Cypress.$(window.top.document.body)).within(() => {
         return cy
@@ -82,6 +79,18 @@ describe('Svelte mount', () => {
         .click()
         .within(() => cy.get('.command-name-mount').should('not.exist'))
       })
+    })
+  })
+
+  it('throws error when receiving removed mounting option', () => {
+    Cypress.on('fail', (e) => {
+      expect(e.message).to.contain('The `styles` mounting option is no longer supported.')
+
+      return false
+    })
+
+    cy.mount(Counter, {
+      styles: `body { background: red; }`,
     })
   })
 
@@ -97,11 +106,11 @@ describe('Svelte mount', () => {
 
     it('should remove previous mounted component', () => {
       cy.mount(Counter)
-      cy.contains('h1', 'Count is 0')
-      cy.mount(Counter, { props: { count: 42 } })
-      cy.contains('h1', 'Count is 42')
+      cy.contains('button', 'count is 0')
+      cy.mount(CounterProp, { props: { count: 42 } })
+      cy.contains('button', 'count is 42')
 
-      cy.contains('h1', 'Count is 0').should('not.exist')
+      cy.contains('button', 'count is 0').should('not.exist')
       cy.get('[data-cy-root]').children().should('have.length', 2)
     })
   })